# JpegEncoder - Optimize Some Low Hanging Fruit