--- a/src/ImageSharp/Formats/Jpeg/Components/Block8x8F.cs
+++ b/src/ImageSharp/Formats/Jpeg/Components/Block8x8F.cs
@@ -51,8 +51,10 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Components
         public Vector4 V7R;
 #pragma warning restore SA1600 // ElementsMustBeDocumented
 
+#if SUPPORTS_RUNTIME_INTRINSICS
         private static readonly Vector<float> NegativeOneAvx = new Vector<float>(-1F);
         private static readonly Vector<float> OffsetAxv = new Vector<float>(.5F);
+#endif
         private static readonly Vector4 NegativeOne = new Vector4(-1);
         private static readonly Vector4 Offset = new Vector4(.5F);
 
@@ -608,7 +610,7 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Components
         }
 
 #if SUPPORTS_RUNTIME_INTRINSICS
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        // [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private static Vector256<float> DivideRoundAvx(
             ref Vector4 dividend,
             ref Vector4 divisor)