… src/ImageSharp/Formats/Jpeg/Components/RowOctet.cs
… src/ImageSharp/Formats/Jpeg/JpegEncoderCore.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/EncodeJpeg.cs
… tests/ImageSharp.Tests/Formats/Jpg/GenericBlock8x8Tests.cs
… tests/ImageSharp.Tests/Image/ImageTests.WrapMemory.cs
… tests/ImageSharp.Tests/TestUtilities/TestEnvironment.cs