### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
<!-- A description of the changes proposed in the pull-request -->
I had a go at optimizing some of the methods identified by @tkp1n here
https://github.com/SixLabors/ImageSharp/issues/1476#issuecomment-764876070

BEFORE

```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-8650U CPU 1.90GHz (Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.102
  [Host]     : .NET Core 5.0.2 (CoreCLR 5.0.220.61120, CoreFX 5.0.220.61120), X64 RyuJIT
  DefaultJob : .NET Core 5.0.2 (CoreCLR 5.0.220.61120, CoreFX 5.0.220.61120), X64 RyuJIT
```

|                Method |     Mean |    Error |   StdDev | Ratio |
|---------------------- |---------:|---------:|---------:|------:|
| 'System.Drawing Jpeg' | 17.37 ms | 0.092 ms | 0.082 ms |  1.00 |
|     'ImageSharp Jpeg' | 23.58 ms | 0.220 ms | 0.205 ms |  1.36 |

AFTER

```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-8650U CPU 1.90GHz (Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.102
  [Host]     : .NET Core 5.0.2 (CoreCLR 5.0.220.61120, CoreFX 5.0.220.61120), X64 RyuJIT
  DefaultJob : .NET Core 5.0.2 (CoreCLR 5.0.220.61120, CoreFX 5.0.220.61120), X64 RyuJIT
```


|                Method |     Mean |    Error |   StdDev | Ratio | RatioSD |
|---------------------- |---------:|---------:|---------:|------:|--------:|
| 'System.Drawing Jpeg' | 17.98 ms | 0.149 ms | 0.132 ms |  1.00 |    0.00 |
|     'ImageSharp Jpeg' | 21.33 ms | 0.302 ms | 0.296 ms |  1.19 |    0.02 |

<!-- Thanks for contributing to ImageSharp! -->