@JimBobSquarePants 2021-02-03
This is all working on my SB2. I'll have to crack open the mac to see what's happening here. 😢