@JimBobSquarePants 2021-02-03
Definitely a Jit issue. Removing inlining changed the error. I’ll manually inline everything into the method.