@JimBobSquarePants 2021-02-03
Everything works now. Will try and recreate the JIT issue and open an issue upstream