@saucecontrol 2021-02-03
Is this code just for dividing your 8x8 blocks by the quantization table?  If so, why not calculate the reciprocal of the QT values up front so you can multiply instead?  Also, you can probably use `vroundps` (`Avx.RoundToNearestInteger`) instead of the offset calc you currently have, assuming you're just truncating that result later.