@JimBobSquarePants 2021-02-04
>Is this code just for dividing your 8x8 blocks by the quantization table? If so, why not calculate the reciprocal of the QT values up front so you can multiply instead? Also, you can probably use vroundps (Avx.RoundToNearestInteger) instead of the offset calc you currently have, assuming you're just truncating that result later.

I'll be honest @saucecontrol and say it's been a while since I looked at all this so I've forgotten a lot of how jpeg works. I'm just limiting changes to known bottlenecks (_that part was actually the least of the three_).

Regarding `Avx.RoundToNearestInteger`. I thought best to leave the expected result for both pipelines identical for now.