@saucecontrol 2021-02-04
Ha, I've just started re-learning everything I've ever semi-understood about JPEG.  Didn't expect the quantization to be much of a bottleneck, but it seems you got some good gains from widening the vectors, so it may be worth looking into some other time.