@JimBobSquarePants 2021-02-04
>Ha, I've just started re-learning everything I've ever semi-understood about JPEG

Man... I cannot wait to see what you bring to the table.