@JimBobSquarePants 2021-02-04
Getting so close now...

```
BenchmarkDotNet=v0.12.1.1467-nightly, OS=Windows 10.0.19042
Intel Core i7-8650U CPU 1.90GHz (Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
.NET SDK=5.0.102
  [Host]       : .NET 5.0.2 (5.0.220.61120), X64 RyuJIT
  .Net 5.0 CLI : .NET 5.0.2 (5.0.220.61120), X64 RyuJIT

Job=.Net 5.0 CLI  Arguments=/p:DebugType=portable  Toolchain=.NET 5.0
IterationCount=5  LaunchCount=1  WarmupCount=5
```

|                                Method |      Mean |     Error |   StdDev | Ratio |     Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|-------------------------------------- |----------:|----------:|---------:|------:|----------:|----------:|----------:|----------:|
|   'System.Drawing Load, Resize, Save' | 410.11 ms | 20.926 ms | 3.238 ms |  1.00 |         - |         - |         - |     11 KB |
|       'ImageSharp Load, Resize, Save' | 182.42 ms |  9.192 ms | 2.387 ms |  0.45 |  333.3333 |         - |         - |  1,988 KB |
|      'ImageMagick Load, Resize, Save' | 448.33 ms | 24.510 ms | 3.793 ms |  1.09 |         - |         - |         - |     53 KB |
|        'ImageFree Load, Resize, Save' | 274.78 ms | 14.070 ms | 2.177 ms |  0.67 | 6000.0000 | 6000.0000 | 6000.0000 |     91 KB |
|      'MagicScaler Load, Resize, Save' |  81.91 ms |  6.855 ms | 1.780 ms |  0.20 |         - |         - |         - |    103 KB |
| 'SkiaSharp Canvas Load, Resize, Save' | 268.31 ms | 26.330 ms | 6.838 ms |  0.65 |         - |         - |         - |     99 KB |
| 'SkiaSharp Bitmap Load, Resize, Save' | 260.10 ms |  9.217 ms | 2.394 ms |  0.64 |         - |         - |         - |     84 KB |
|          'NetVips Load, Resize, Save' | 162.13 ms | 13.306 ms | 3.456 ms |  0.39 |         - |         - |         - |     47 KB |